requirements.txt
aiohttp==3.7.4.post0
anyio==3.2.1
astroid==2.6.2
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
buttons==0.1.9
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
discord-components==1.1.2
discord-py-slash-command==2.3.1
discord.py==1.7.3
h11==0.12.0
httpcore==0.13.6
httpx==0.18.2
idna==2.10
isort==5.9.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
multidict==5.1.0
pistonapi==1.0.1
psutil==5.8.0
pycodestyle==2.7.0
pycparser==2.20
pylint==2.9.3
PyNaCl==1.4.0
python-dotenv==0.18.0
requests==2.25.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
wrapt==1.12.1
yarl==1.6.3
youtube-dl==2021.6.6
youtube-search-python==1.4.6